requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
pdfplumber==0.11.0
python-docx==1.1.2
nltk==3.8.1
reportlab==4.2.0
Werkzeug==3.0.3
requests==2.29.0
python-dotenv==1.0.1
Pillow==10.3.0
gunicorn==22.0.0
Flask-CORS==4.0.1
PyMuPDF==1.24.3